Add pagination_class and move helper to utils
Add pagination_class and move helper to utils
