+++ utils/paginations.py
@@ -1,0 +1,48 @@
+from rest_framework.pagination import BasePagination, PageNumberPagination
+from rest_framework.response import Response
+
+
+class EndlessPagination(BasePagination):
+    page_size = 20
+
+    def __init__(self):
+        super(EndlessPagination, self).__init__()
+        self.has_next_page = False
+
+    def to_html(self):
+        pass
+
+    def paginate_queryset(self, queryset, request, view=None):
+        if 'created_at__gt' in request.query_params:
+            created_at__gt = request.query_params['created_at__gt']
+            queryset = queryset.filter(id__gt=created_at__gt)
+            self.has_next_page = False
+            return queryset.order_by('id')
+
+        if 'created_at__lt' in request.query_params:
+            created_at__lt = request.query_params['created_at__lt']
+            queryset = queryset.filter(id__lt=created_at__lt)
+
+        queryset = queryset.order_by('-id')[:self.page_size + 1]
+        self.has_next_page = len(queryset) > self.page_size
+        return queryset[:self.page_size]
+
+    def get_paginated_response(self, data):
+        return Response({
+            'has_next_page': self.has_next_page,
+            'results': data,
+        })
+
+class FriendshipPagination(PageNumberPagination):
+    page_size = 20
+    page_size_query_param = 'size'
+    max_page_size = 20
+
+    def get_paginated_response(self, data):
+        return Response({
+            'total_results': self.page.paginator.count,
+            'total_pages': self.page.paginator.number_pages,
+            'page_number': self.page.number,
+            'has_next_page': self.page.has_next(),
+            'results': data,
+        })
